refactor(ios): refactor touches event for view
no HippyTouchHandler any more, touches events dont depend on HippyRootView
