+++ ios/sdk/base/HippyBridge.h
@@ -28,7 +28,6 @@
 #import "HippyFrameUpdate.h"
 #import "HippyInvalidating.h"
 #import "HippyImageViewCustomLoader.h"
-#import "HippyCustomTouchHandlerProtocol.h"
 #import "HippyImageProviderProtocol.h"
 
 @class JSValue;
@@ -205,7 +204,6 @@ HIPPY_EXTERN NSString *HippyBridgeModuleNameForClass(Class bridgeModuleClass);
 @property (nonatomic, weak, readonly) HippyExtAnimationModule *animationModule;
 
 @property (nonatomic, strong, readonly) id<HippyImageViewCustomLoader> imageLoader;
-@property (nonatomic, strong, readonly) id<HippyCustomTouchHandlerProtocol> customTouchHandler;
 @property (nonatomic, strong, readonly) NSSet<Class<HippyImageProviderProtocol>> *imageProviders;
 
 /**
